Remove AgentChatPanel implementation file
Delete ui-crates/ui_core/src/app/agent_chat_panel.rs which contained the full AgentChatPanel UI implementation (~2.7k LOC). Update ui-crates/ui_core/src/app/agent_chat_panel/mod.rs to reflect the removal (module reorganization / consolidation). This cleans up the module layout and centralizes the panel's public surface in the mod file.
